backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# ML/DRL
torch>=2.7.0
numpy>=1.26.0

# Visualization
matplotlib>=3.8.0

# Utilities (optional)
pandas>=2.1.0
